% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assays}
\alias{assays}
\title{Assays for mutation detection using different primers and probes}
\format{
A dataframe with 10 rows and 7 columns.
}
\source{
This data is created by Zhanglab in Xiamen University.
}
\usage{
data(assays)
}
\description{
These assays include the primer sequences and the detection ranges in which
mutations may occur.
}
\examples{
data(assays)
}
\keyword{datasets}
