% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinalist}
\alias{chinalist}
\title{A list of places in China}
\format{
A dataframe with 31 rows and 1 column.
}
\source{
This data is created by Zhanglab in Xiamen University.
}
\usage{
data(chinalist)
}
\description{
The list is used for displacing some original cities' names with "China" in order
to make the downstream analysis easier.
}
\examples{
data(chinalist)
}
\keyword{datasets}
