% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_annot}
\alias{covid_annot}
\title{Mutation annotation results produced by "indelSNP" function}
\format{
A large matrix.
}
\source{
\url{https://www.gisaid.org/}
}
\usage{
data(covid_annot)
}
\description{
A large matrix which could be used for downstream analysis like mutation
statistics description.
}
\examples{
data(covid_annot)
}
\keyword{datasets}
