% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_position}
\alias{gene_position}
\title{"GFF3" format gene position data for SARS-Cov-2}
\format{A dataframe with 26 rows and 10 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/}
}
\usage{
data(gene_position)
}
\description{
This "GFF3" data is used for counting the mutations in each gene in virus sample.
}
\examples{
data(gene_position)
}
\keyword{datasets}
