% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gff3}
\alias{gff3}
\title{"GFF3" format annotation data for SARS-Cov-2}
\format{
A dataframe with 26 rows and 10 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi?id=2697049}
}
\usage{
data(gff3)
}
\description{
This "GFF3" data is used for annotating the effects of mutations in virus sample.
}
\examples{
data(gff3)
}
\keyword{datasets}
