% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nucmerr}
\alias{nucmerr}
\title{Preprocessed "nucmer.snpss" file using "nucmerRMD" function}
\format{A dataframe with 150000 rows(downsampled mutation sites) and 10
columns.}
\source{
\url{https://www.gisaid.org/}
}
\usage{
data(nucmerr)
}
\description{
A dataset contains some group information subtracted from the "nucmer"
object by "nucmerRMD" function in order to best describe the results.
}
\examples{
data(nucmerr)
}
\keyword{datasets}
