% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refseq}
\alias{refseq}
\title{SARS-Cov-2 genomic reference sequence from NCBI}
\format{
"SeqFastadna" characters.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/32015508/}
}
\usage{
data(refseq)
}
\description{
This reference sequence is derived from "fasta" file, preprocessed by
"read.fasta" function(refseq<-read.fasta("NC_045512.2.fa",forceDNAtolower=FALSE)[[1]]). It is used for annotating mutations in virus samples.
}
\examples{
data(refseq)
}
\keyword{datasets}
