% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assays}
\alias{assays}
\title{Assays for mutation detection using different primers and probes}
\format{
A dataframe with 10 rows and 7 columns.
}
\usage{
data(assays)
}
\description{
These assays include the primer detection ranges in which
mutations may occur.
}
\examples{
data(assays)
}
\references{
Kilic T, Weissleder R, Lee H (2019) iScience 23, 101406.
(\href{https://pubmed.ncbi.nlm.nih.gov/32771976/}{PubMed})
}
\keyword{datasets}
