% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_annot}
\alias{covid_annot}
\title{Mutation annotation results produced by "indelSNP" function}
\format{
A dataframe with 49821 rows and 10 columns.
}
\source{
\url{https://www.gisaid.org/}
}
\usage{
data(covid_annot)
}
\description{
A dataframe which could be used for downstream analysis like mutation
statistics description.
}
\examples{
data(covid_annot)
}
\keyword{datasets}
