% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nucmer}
\alias{nucmer}
\title{Mutation information derived from "nucmer" SNP analysis}
\format{
A dataframe with 5000 rows(mutation sites) and 14 columns.
}
\source{
\url{https://www.gisaid.org/}
}
\usage{
data(nucmer)
}
\description{
The "nucmer.snpss" variant file is obtained by processing the SARS-Cov-2 sequence
from Gisaid website (complete, high coverage only, low coverage exclusion, Host=human,
Virus name = hCoV-19) with "seqkit" software and "nucmer" scripts. The example data
is downsampled from complete data in 2020-06-14.
}
\examples{
data(nucmer)
}
\keyword{datasets}
